+++ firebase/functions/src/quarters.ts
@@ -1,26 +1,6 @@
 import {Mutex} from "async-mutex";
 import dayjs from "dayjs";
-
-export interface Quarter {
-  "quarter": string;
-  "qyy": string;
-  "name": string;
-  "category": string;
-  "academicYear": string;
-  "firstDayOfClasses": string;
-  "lastDayOfClasses": string;
-  "firstDayOfFinals": string;
-  "lastDayOfFinals": string;
-  "firstDayOfQuarter": string;
-  "lastDayOfSchedule": string;
-  "pass1Begin": string;
-  "pass2Begin": string;
-  "pass3Begin": string;
-  "feeDeadline": string;
-  "lastDayToAddUnderGrad": string;
-  "lastDayToAddGrad": string;
-  "lastDayThirdWeek": string;
-}
+import {Quarter} from "../../../helpers/api";
 
 const quartersMutex = new Mutex();
 let quarters: Record<string, Quarter> | null = null;
