+++ types/index.d.ts
@@ -1,0 +1,3 @@
+export * from "./canvas";
+export * from "./ucsb";
+export * from "./quarters";
