+++ types/quarters.d.ts
@@ -1,0 +1,20 @@
+export interface Quarter {
+  "quarter": string;
+  "qyy": string;
+  "name": string;
+  "category": string;
+  "academicYear": string;
+  "firstDayOfClasses": string;
+  "lastDayOfClasses": string;
+  "firstDayOfFinals": string;
+  "lastDayOfFinals": string;
+  "firstDayOfQuarter": string;
+  "lastDayOfSchedule": string;
+  "pass1Begin": string;
+  "pass2Begin": string;
+  "pass3Begin": string;
+  "feeDeadline": string;
+  "lastDayToAddUnderGrad": string;
+  "lastDayToAddGrad": string;
+  "lastDayThirdWeek": string;
+}
